python01/ex05/requirements.txt
contourpy==1.3.2
cycler==0.12.1
flake8==7.3.0
fonttools==4.61.0
kiwisolver==1.4.9
matplotlib==3.10.7
mccabe==0.7.0
numpy==2.2.6
packaging==25.0
pillow==12.0.0
pycodestyle==2.14.0
pyflakes==3.4.0
pyparsing==3.2.5
PyQt6==6.10.0
PyQt6-Qt6==6.10.1
PyQt6_sip==13.10.2
python-dateutil==2.9.0.post0
six==1.17.0
